Refactor to move to inside App folder
