+++ src/App/NewGameMenu/GenerateCode.tsx
