+++ src/App/NewGameMenu/NewGameMenu.stories.tsx
